ai-orchestrator/requirements.txt
pyyaml>=6.0
python-dotenv>=1.0.0

# Optional - install what you need
openai>=1.0.0
anthropic>=0.18.0
requests>=2.31.0  # for Ollama

# Dev
pytest>=7.4.0
black>=23.0.0
mypy>=1.5.0
